requirements.txt
wrds
pandas
numpy
requests
python-dotenv
feedparser
beautifulsoup4
lxml
